+++ metadata.rb
@@ -1,7 +1,7 @@
-name             'Site24x7'
-maintainer       "Opscode, Inc."
-maintainer_email "sysadmin@opscode.com"
+name             'site24x7'
+maintainer       'Zoho corporation'
+maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures Site24x7'
+description      'Installs/Configures site24x7'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'